backend/requirements.txt
pandas
openai-whisper
librosa
numpy
torch
faiss-cpu
transformers
langchain
langchain-community
langchain-huggingface
langchain-google-genai
sentence-transformers
watchdog
beautifulsoup4
requests
pymupdf
trafilatura
chardet
flashrank
flask
flask-cors
pymupdf